wip
2 'bugs' might have been fixed btw
