--- .eslintrc
@@ -1,38 +1,0 @@
-{
-    "parser"  : "babel-eslint",
-    "extends" : [
-        "standard",
-        "standard-react"
-     ],
-     "plugins": [
-         "react"
-     ],
-    "env" : {
-        "browser" : true,
-        "es6": true
-    },
-    "globals": {
-        "__DEV__": false
-    },
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "ecmaFeatures": {
-            "jsx": true
-        }
-    },
-    "rules": {
-        "generator-star-spacing": 0,
-        "indent": [2, 4, { "ignoredNodes": ["JSXAttribute", "JSXSpreadAttribute"], "SwitchCase": 1 }],
-        "no-warning-comments": [1, {
-            "terms": ["todo", "fixme", "xxx"],
-            "location": "start"
-        }],
-        "operator-linebreak": [2, "after"],
-        "padded-blocks": 0,
-        "semi": [2, "always"],
-        "react/jsx-indent-props": [2, 2],
-        "react/jsx-boolean-value": [0, "never"],
-        "react/jsx-curly-spacing": [0, "never"],
-        "react/jsx-indent": [2, 4]
-    }
-}
