+++ .gitignore
@@ -3,3 +3,5 @@ node_modules
 #LOCK
 package-lock.json
 yarn.lock
+
+lib
