+++ package.json
@@ -2,10 +2,18 @@
     "name": "react-native-snap-carousel",
     "version": "4.0.0-beta.5",
     "description": "Swiper/carousel component for React Native with previews, multiple layouts, parallax images, performant handling of huge numbers of items, and more. Compatible with Android & iOS.",
-    "main": "src/index.js",
+    "main": "lib/commonjs/index",
+    "module": "lib/module/index",
+    "types": "lib/typescript/src/index.d.ts",
+    "react-native": "src/index",
+    "source": "src/index",
     "repository": {
         "type": "git",
         "url": "github.com/archriss/react-native-snap-carousel"
+    },
+    "scripts": {
+        "prepare": "bob build",
+        "lint": "eslint \"**/*.{js,ts,tsx}\""
     },
     "keywords": [
         "react",
@@ -36,6 +44,7 @@
     "author": "Benoit Delmaire <bdelmaire@archriss.com> (github.com/bd-arc)",
     "license": "BSD-3-Clause",
     "dependencies": {
+        "@types/react-addons-shallow-compare": "^0.14.22",
         "prop-types": "15.7.2",
         "react-addons-shallow-compare": "15.6.2"
     },
@@ -44,15 +53,55 @@
         "react-native": ">=0.58.0"
     },
     "devDependencies": {
+        "@react-native-community/bob": "^0.16.2",
+        "@react-native-community/eslint-config": "^2.0.0",
+        "@types/react": "^16.9.19",
+        "@types/react-native": "0.62.13",
         "babel-eslint": "10.1.0",
-        "eslint": "6.8.0",
-        "eslint-config-standard": "14.1.1",
-        "eslint-config-standard-react": "9.2.0",
-        "eslint-plugin-import": "2.20.2",
-        "eslint-plugin-node": "11.1.0",
-        "eslint-plugin-promise": "4.2.1",
-        "eslint-plugin-react": "7.19.0",
-        "eslint-plugin-standard": "4.0.1"
+        "eslint": "^7.2.0",
+        "eslint-config-prettier": "^6.11.0",
+        "eslint-plugin-prettier": "^3.1.3",
+        "react-native": "0.61.5",
+        "typescript": "^3.8.3"
+    },
+    "eslintConfig": {
+        "extends": [
+          "@react-native-community",
+          "prettier"
+        ],
+        "rules": {
+          "prettier/prettier": [
+            "error",
+            {
+              "quoteProps": "consistent",
+              "singleQuote": true,
+              "tabWidth": 2,
+              "trailingComma": "es5",
+              "useTabs": false
+            }
+          ]
+        }
+      },
+    "eslintIgnore": [
+        "node_modules/",
+        "lib/",
+        "example"
+    ],
+    "prettier": {
+        "quoteProps": "consistent",
+        "singleQuote": true,
+        "tabWidth": 2,
+        "trailingComma": "es5",
+        "useTabs": false
+    },
+    "@react-native-community/bob": {
+        "source": "src",
+        "output": "lib",
+        "targets": [
+            "commonjs",
+            "module",
+            "typescript"
+        ]
     },
     "homepage": "https://github.com/archriss/react-native-snap-carousel",
     "bugs": {
