+++ src/index.ts
@@ -3,4 +3,10 @@ import Pagination from './pagination/Pagination';
 import ParallaxImage from './parallaximage/ParallaxImage';
 import { getInputRangeFromIndexes } from './utils/animations';
 
-export { Carousel as default, Pagination, ParallaxImage, getInputRangeFromIndexes };
+export {
+  Carousel as default,
+  Carousel,
+  Pagination,
+  ParallaxImage,
+  getInputRangeFromIndexes,
+};
